% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UK_money_format.R
\name{UK_money_format}
\alias{UK_money_format}
\title{UK Money Format}
\usage{
UK_money_format(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector.}

\item{digits}{Rounding in \code{abs(x) / scale}, defaults to 0.}
}
\value{
A character vector of length corresponding to the numeric vector.
  \code{NA} values will be represented as \code{"£NA"}.
}
\details{
\code{UK_money_format} takes numeric vector and returns character vector with
  of corresponding length with values reflecting common way of expressing UK currency.
  Consequently, 1e6 becomes £1m. This is fairly arbitrary but may
  be useful when applied in generating chart labels. For alternative approaches,
  use of suffixes and units see: \url{https://stackoverflow.com/q/28159936/1655567}.
}
\examples{
UK_money_format(c(9, -10, -1234765.12, NA))
# [1] "£9"   "-£10" "-£1m" NA
\dontrun{
UK_money_format("a")
# Error: is.numeric(x) is not TRUE
}
}
